requirements.txt
Django==1.1.1
django-reversion==1.2.1
wsgiref==0.1.2
